codex: carve threads.rs (pure move)
Thread actions (compact/fork/side threads, rollback, review, goals,
naming) -- the first inherent impl block moves whole plus the
developer-instruction and goal-formatting helpers it leans on; 27
tests and their worktree fixture helper ride along, the shared
test_agent fixture stays in mod.rs pub(crate)-widened.
